refactor: moved tests around to their module specific files
