+++ backend/src/test/room/test_waypoint_pojo.py
@@ -45,3 +45,45 @@ def test_has_entrance_a_on_burrow_returns_false(waypoint_pojo_a):
     actual = waypoint_pojo_a.has_entrance(room)
     expected = False
     assert actual == expected
+
+
+def test_is_entrance_a_on_tile_returns_true(waypoint_pojo_a):
+    tile = Tile(0, 0, DungeonCardValues.ENTRANCE_A)
+    actual = waypoint_pojo_a.is_entrance(tile)
+    expected = True
+    assert actual == expected
+
+
+def test_is_entrance_a_on_non_entrance_return_false(waypoint_pojo_a):
+    tile = Tile(0, 0, DungeonCardValues.ENTRANCE_B)
+    actual = waypoint_pojo_a.is_entrance(tile)
+    expected = False
+    assert actual == expected
+
+
+def test_is_entrance_b_on_non_entrance_return_false(waypoint_pojo_b):
+    tile = Tile(0, 0, DungeonCardValues.ENTRANCE_A)
+    actual = waypoint_pojo_b.is_entrance(tile)
+    expected = False
+    assert actual == expected
+
+
+def test_has_entrance_b_on_trail_returns_false(waypoint_pojo_b):
+    room = Trail()
+    actual = waypoint_pojo_b.has_entrance(room)
+    expected = False
+    assert actual == expected
+
+
+def test_get_entrance_a_on_trail_returns_entrance_tile(waypoint_pojo_a):
+    room = Trail()
+    actual = waypoint_pojo_a.get_entrance(room)
+    expected = Tile(-2, 5, DungeonCardValues.ENTRANCE_A)
+    assert actual == expected
+
+
+def test_get_entrance_b_on_hovel_returns_entrance_tile(waypoint_pojo_b):
+    room = Hovel()
+    actual = waypoint_pojo_b.get_entrance(room)
+    expected = Tile(0, 3, DungeonCardValues.ENTRANCE_B)
+    assert actual == expected
